+++ src/test/java/com/digitalkitchen/meals/util/TestConstants.java
@@ -2,7 +2,7 @@ package com.digitalkitchen.meals.util;
 
 public class TestConstants {
 
-    public static int RECIPE_ID = 1337;
+    public static long RECIPE_ID = 1337;
     public static long MEAL_ID = 555;
     public static long MEAL_PLAN_ID = 444;
     public static long MEAL_RECIPE_ID = 777;
